fix:チャネルトークの出し分けをmixinで共通化
