+++ channel_talk_test/src/pages/Contact.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import ChannelService from '../channelTalk/channelTalk'
+import {Mixin} from '../js/mixin'
 
 export default {
   data() {
@@ -14,15 +14,8 @@ export default {
       route: this.$route,
     };
   },
-  created: function() {
-    if (this.route.path == '/contact') {
-        ChannelService.boot({
-          "pluginKey": "3ad99376-fc05-4f64-862c-9b8772628dd6",
-        });
-    }
-  }
+  mixins:[Mixin]
 };
-
 </script>
 
 <style scoped>
